Move index/home page to sub folder
This makes sure, that the loading state only applies for the index page
